+++ packages/statsd/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- Update `hot-spot` dependencies [[#1379](https://github.com/Shopify/quilt/pull/1379)]
+
+### Added
+
+- added `timing` metric support from `hot-spot`
 
 ## [1.2.0] - 2019-12-24